Lucas/Robo/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
schedule==1.2.1
loguru==0.7.2
PyQt6==6.4.2
